+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ~~~~~~~~~~~~~~~~~~
 
 -  Switch to MyST-Parser from recommonmark.
+-  Drop support for Sphinx directives.
 
 0.0.26 (2021-02-13)
 ~~~~~~~~~~~~~~~~~~~